+++ core/forms.py
@@ -1,7 +1,11 @@
 from django import forms
-from .models import Aluno
 
-class UploadFileExcelForm(forms.Form):
-    title = forms.CharField(max_length=50)
-    file = forms.FileField()
+class DadosExcelForm(forms.Form):
+
+    title = forms.CharField(max_length=50, widget=forms.TextInput(
+        attrs={'class': 'file-path validate'})
+    )
+    file = forms.FileField(widget=forms.TextInput(
+        attrs={'type': 'file', 'accept': 'application/vnd.ms-excel, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'})
+    )
